pyproject.toml
[tool.poetry]
name = "telegram-audit-bot"
version = "0.1.0"
description = "Telegram бот для анализа и аудита веб-сайтов"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.7"
httpx = "^0.25.2"
beautifulsoup4 = "^4.12.2"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
loguru = "^0.7.2"
tenacity = "^8.2.3"
gspread = "^5.12.0"
google-auth = "^2.25.2"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
uvloop = {version = "^0.19.0", optional = true}
openai = "^1.107.1"

[tool.poetry.scripts]
run = "app.telegram.bot:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
ruff = "^0.1.7"
black = "^23.11.0"

[tool.poetry.extras]
uvloop = ["uvloop"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"